data/pyproject.toml
[project]
name = "data"
version = "0.1.0"
description = "Add your description here"
readme = "../README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "numpy>=2.4.2",
    "pandas>=3.0.1",
    "progressbar2>=4.5.0",
]
